Added documentation and improved code quality
