+++ .gitignore
@@ -1,3 +1,5 @@
 target
 Cargo.lock
 .env
+.idea
+.DS_Store
