--- .idea/shelf/operators_single_rs.xml
@@ -1,4 +1,0 @@
-<changelist name="operators_single_rs" date="1703524251309" recycled="false">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/operators_single_rs/shelved.patch" />
-  <option name="DESCRIPTION" value="operators_single.rs" />
-</changelist>
