+++ examples/axum-echo/axum_echo.rs
@@ -1,9 +1,9 @@
 use axum::routing::get;
 use serde_json::Value;
-use socketioxide::{extract::{AckSender, Bin, Data, SocketRef}, SocketIo, SocketIoBuilder};
+use socketioxide::{extract::{AckSender, Bin, Data, SocketRef}, SocketIoBuilder};
 use tracing::info;
 use tracing_subscriber::FmtSubscriber;
-use socketioxide::parser::MsgpackParser;
+use socketioxide::parser::msgpack::MsgpackParser;
 
 fn on_connect(socket: SocketRef, Data(data): Data<Value>) {
     info!("Socket.IO connected: {:?} {:?}", socket.ns(), socket.id);
@@ -31,7 +31,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     tracing::subscriber::set_global_default(FmtSubscriber::default())?;
 
     let (layer, io) = SocketIoBuilder::new()
-        .with_parser(MsgpackParser::default())
+        .with_parser(MsgpackParser)
         .build_layer();
 
     io.ns("/", on_connect);
